requirements.txt
accelerate==1.0.1
certifi==2025.6.15
charset-normalizer==3.4.2
diffusers==0.34.0
filelock==3.16.1
fsspec==2025.3.0
hf-xet==1.1.5
huggingface-hub==0.33.0
idna==3.10
importlib-metadata==8.5.0
jinja2==3.1.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
packaging==25.0
pillow==10.4.0
psutil==7.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
sympy==1.13.3
tokenizers==0.20.3
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.46.3
typing-extensions==4.13.2
urllib3==2.2.3
zipp==3.20.2
fastapi==0.104.1
uvicorn==0.24.0
